feat: add `appendHostToAmpUrl` to make AMPUrl (`rel="amphtml"`) absolute.
Based on https://acceleratedmobilepageurl.googleapis.com/v1/ampUrls:batchGet api response, if rel="amphtml" is not an absolute URL, google seems to be considering it an Error with response:

{
  "urlErrors": [
    {
      "errorCode": "URL_IS_INVALID_AMP",
      "errorMessage": "Request URL is an invalid AMP URL.",
      "originalUrl": "https://lonely-part.surge.sh"
    }
  ]
}
I didn't move every AMP to absolute in this PR, but made it configurable, so that it will be easy to move. Once this hypothesis is proved we can remove the toggle altogether also.